Refactor, flytt felles logikk inn i toggle som henter alt, men wrappere lever videre
