+++ apps/etterlatte-saksbehandling-ui/client/src/App.tsx
@@ -28,8 +28,6 @@ import { isSuccess } from '~shared/api/apiUtils'
 import { isFailureHandler } from '~shared/api/IsFailureHandler'
 import { FlyttJournalpost } from '~components/person/flyttjournalpost/FlyttJournalpost'
 import { FEATURE_TOGGLE_KAN_BRUKE_KLAGE } from '~components/person/KlageListe'
-import { MinOppgaveliste } from '~components/oppgavebenk/MinOppgaveliste'
-import { OppgavelistaWrapper } from '~components/oppgavebenk/OppgavelistaWrapper'
 
 function App() {
   const innloggetbrukerHentet = useInnloggetSaksbehandler()
@@ -57,8 +55,7 @@ function App() {
               <ConfigContext.Provider value={hentConfigStatus.data}>
                 <Routes>
                   <Route path="/" element={<ToggleMinOppgaveliste />}>
-                    <Route index element={<OppgavelistaWrapper />} />
-                    <Route path="/minoppgaveliste" element={<MinOppgaveliste />} />
+                    <Route path="/minoppgaveliste" element={<ToggleMinOppgaveliste />} />
                   </Route>
 
                   <Route path="/flyttjournalpost" element={<FlyttJournalpost />} />
