+++ apps/etterlatte-saksbehandling-ui/client/src/components/oppgavebenk/ToggleMinOppgaveliste.tsx
@@ -2,10 +2,19 @@ import React, { useEffect, useState } from 'react'
 import { Tabs } from '@navikt/ds-react'
 import { InboxIcon, PersonIcon } from '@navikt/aksel-icons'
 import styled from 'styled-components'
-import { useAppSelector } from '~store/Store'
+import { useAppDispatch, useAppSelector } from '~store/Store'
 import { Container } from '~shared/styled'
 import { Tilgangsmelding } from '~components/oppgavebenk/Tilgangsmelding'
-import { Outlet, useLocation, useNavigate } from 'react-router-dom'
+import { useLocation, useNavigate } from 'react-router-dom'
+import { Filter, minOppgavelisteFiltre } from '~components/oppgavebenk/filter/oppgavelistafiltre'
+import { hentFilterFraLocalStorage, leggFilterILocalStorage } from '~components/oppgavebenk/filter/filterLocalStorage'
+import { useApiCall } from '~shared/hooks/useApiCall'
+import { hentGosysOppgaver, hentOppgaverMedStatus, OppgaveDTO } from '~shared/api/oppgaver'
+import { isSuccess } from '~shared/api/apiUtils'
+import { sorterOppgaverEtterOpprettet } from '~components/oppgavebenk/oppgaveutils'
+import { settHovedOppgavelisteLengde, settMinOppgavelisteLengde } from '~store/reducers/OppgavelisteReducer'
+import { MinOppgaveliste } from '~components/oppgavebenk/MinOppgaveliste'
+import { OppgavelistaWrapper } from '~components/oppgavebenk/OppgavelistaWrapper'
 
 type OppgavelisteToggle = 'Oppgavelista' | 'MinOppgaveliste'
 
@@ -37,6 +46,92 @@ export const ToggleMinOppgaveliste = () => {
     }
   }, [oppgaveListeValg])
 
+  const dispatch = useAppDispatch()
+
+  const [minsideFilter, setMinsideFilter] = useState<Filter>(minOppgavelisteFiltre())
+  const [hovedsideFilter, setHovedsideFilter] = useState<Filter>(hentFilterFraLocalStorage())
+
+  useEffect(() => {
+    hentMinsideOppgaver(minsideFilter.oppgavestatusFilter)
+  }, [minsideFilter.oppgavestatusFilter])
+
+  useEffect(() => {
+    leggFilterILocalStorage(hovedsideFilter)
+  }, [hovedsideFilter])
+
+  const [minsideOppgaverResult, hentOppgaverMinside] = useApiCall(hentOppgaverMedStatus)
+  const [hovedsideOppgaverResult, hentAlleOppgaverStatusFetch] = useApiCall(hentOppgaverMedStatus)
+  const [gosysOppgaverResult, hentGosysOppgaverFunc] = useApiCall(hentGosysOppgaver)
+
+  const hentMinsideOppgaver = (oppgavestatusFilter: Array<string> | undefined) =>
+    hentOppgaverMinside({
+      oppgavestatusFilter: oppgavestatusFilter ? oppgavestatusFilter : minsideFilter.oppgavestatusFilter,
+      minOppgavelisteIdent: true,
+    })
+  const hentHovedsideOppgaver = (oppgavestatusFilter: Array<string> | undefined) =>
+    hentAlleOppgaverStatusFetch({
+      oppgavestatusFilter: oppgavestatusFilter ? oppgavestatusFilter : hovedsideFilter.oppgavestatusFilter,
+      minOppgavelisteIdent: false,
+    })
+
+  const hentHovedsideOppgaverAlle = () => {
+    hentMinsideOppgaver(undefined)
+    hentGosysOppgaverFunc({})
+  }
+
+  const hentAlleOppgaver = () => {
+    hentMinsideOppgaver(undefined)
+    hentHovedsideOppgaver(undefined)
+    hentGosysOppgaverFunc({})
+  }
+
+  const filtrerKunInnloggetBrukerOppgaver = (oppgaver: Array<OppgaveDTO>) => {
+    return oppgaver.filter((o) => o.saksbehandlerIdent === innloggetSaksbehandler.ident)
+  }
+
+  const [hovedsideOppgaver, setHovedsideOppgaver] = useState<Array<OppgaveDTO>>([])
+  const [minsideOppgaver, setMinsideOppgaver] = useState<Array<OppgaveDTO>>([])
+
+  useEffect(() => {
+    hentAlleOppgaver()
+  }, [])
+
+  useEffect(() => {
+    if (isSuccess(hovedsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      const alleOppgaverMerget = sorterOppgaverEtterOpprettet([
+        ...hovedsideOppgaverResult.data,
+        ...gosysOppgaverResult.data,
+      ])
+      setHovedsideOppgaver(alleOppgaverMerget)
+    } else if (isSuccess(hovedsideOppgaverResult) && !isSuccess(gosysOppgaverResult)) {
+      setHovedsideOppgaver(sorterOppgaverEtterOpprettet(hovedsideOppgaverResult.data))
+    } else if (!isSuccess(hovedsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      setHovedsideOppgaver(sorterOppgaverEtterOpprettet(gosysOppgaverResult.data))
+    }
+  }, [hovedsideOppgaverResult, gosysOppgaverResult])
+
+  useEffect(() => {
+    if (isSuccess(minsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      const alleOppgaverMerget = sorterOppgaverEtterOpprettet([
+        ...minsideOppgaverResult.data,
+        ...filtrerKunInnloggetBrukerOppgaver(gosysOppgaverResult.data),
+      ])
+      setMinsideOppgaver(alleOppgaverMerget)
+    } else if (isSuccess(minsideOppgaverResult) && !isSuccess(gosysOppgaverResult)) {
+      setMinsideOppgaver(sorterOppgaverEtterOpprettet(minsideOppgaverResult.data))
+    } else if (!isSuccess(minsideOppgaverResult) && isSuccess(gosysOppgaverResult)) {
+      setMinsideOppgaver(sorterOppgaverEtterOpprettet(filtrerKunInnloggetBrukerOppgaver(gosysOppgaverResult.data)))
+    }
+  }, [gosysOppgaverResult, minsideOppgaverResult])
+
+  useEffect(() => {
+    dispatch(settHovedOppgavelisteLengde(hovedsideOppgaver.length))
+  }, [hovedsideOppgaver])
+
+  useEffect(() => {
+    dispatch(settMinOppgavelisteLengde(minsideOppgaver.length))
+  }, [minsideOppgaver])
+
   return (
     <Container>
       <TabsWidth value={oppgaveListeValg} onChange={(e) => setOppgaveListeValg(e as OppgavelisteToggle)}>
@@ -49,7 +144,27 @@ export const ToggleMinOppgaveliste = () => {
           />
         </Tabs.List>
       </TabsWidth>
-      <Outlet />
+      {oppgaveListeValg === 'MinOppgaveliste' ? (
+        <MinOppgaveliste
+          minsideOppgaver={minsideOppgaver}
+          minsideOppgaverResult={minsideOppgaverResult}
+          gosysOppgaverResult={gosysOppgaverResult} //TODO: hvem skal filtrere for sb sine?
+          minsideFilter={minsideFilter}
+          setMinsideFilter={setMinsideFilter}
+          setMinsideOppgaver={setMinsideOppgaver}
+        />
+      ) : (
+        <OppgavelistaWrapper
+          hovedsideOppgaver={hovedsideOppgaver}
+          hentHovedsideOppgaverAlle={hentHovedsideOppgaverAlle}
+          hovedsideOppgaverResult={hovedsideOppgaverResult}
+          gosysOppgaverResult={gosysOppgaverResult}
+          hentHovedsideOppgaver={hentHovedsideOppgaver}
+          hovedsideFilter={hovedsideFilter}
+          setHovedsideFilter={setHovedsideFilter}
+          setHovedsideOppgaver={setHovedsideOppgaver}
+        />
+      )}
     </Container>
   )
 }
